change Var, connect 2 Anfs

+++ code/tests.c
@@ -398,84 +398,6 @@
 //     return OK;   
 // }
 
-// int anfWithNodesInBundle2() {
-
-//     freopen ("./testOutput/5", "a+", stdout);
-
-//     tANFBundle* bundle = createANFBundle(10, 0.7);
-//     if (bundle == NULL) {
-//         return ERROR;
-//     }
-
-//     tVar varArray[] = {
-//         createVar("var1", true),
-//         createVar("var2", true),
-//         createVar("var3", true),
-//         createVar("var4", true)
-//     };
-
-//     tVar varArray2[] = {
-//         createVar("var11", true),
-//         createVar("var12", true),
-//         createVar("var13", true),
-//         createVar("var14", true)
-//     };
-
-//     tNode* node1 = createNode();
-//     if (node1 == NULL) {
-//         freeANFBundle(bundle);
-//         return ERROR;
-//     }
-
-//     if (addVariablesToNode(varArray, ARRAY_SIZE(varArray), node1) != OK) {
-//         freeNode(node1);
-//         freeANFBundle(bundle);
-//         return ERROR;
-//     } 
-
-//     tNode* node2 = createNode();
-//     if (node2 == NULL) {
-//         freeANFBundle(bundle);
-//         return ERROR;
-//     }
-
-//     if (addVariablesToNode(varArray2, ARRAY_SIZE(varArray2), node2) != OK) {
-//         freeNode(node1);
-//         freeANFBundle(bundle);
-//         return ERROR;
-//     } 
-
-//     printNode(node1);
-//     printNode(node2);
-
-//     tNode* nodes[] = {
-//         node1,
-//         node2
-//     };
-
-//     tAnf* anf1 = createAnfWithNodesInBundle(bundle, nodes, ARRAY_SIZE(nodes));
-//     if (anf1 == NULL) {
-//         freeNode(node1);
-//         freeNode(node2);
-//         freeANFBundle(bundle);
-//         return ERROR;
-//     }
-
-//     printBundle(bundle);
-//     printBundleMap(bundle);
-//     generateAnfGraph(anf1, "./test5.gv");
-
-//     freeNode(node1);
-//     freeNode(node2);
-
-//     freeAnf(anf1);
-//     freeANFBundle(bundle);
-
-//     freopen("/dev/tty", "w", stdout);
-
-//     return OK;   
-// }
-
 // int deleteVarFromNode() {
 
 //     freopen ("./testOutput/6", "a+", stdout);
@@ -664,6 +586,25 @@ int singleFunctionTest(char* graphOutput) {
     printAnf(anf);
     printHashMap(anf->hashMap);
 
+    if (switchVarValueInAnf("var5", anf)) {
+        freeNode(node1);
+        freeNode(node2);
+        freeAnf(anf);
+        return ERROR;    
+    }
+
+    printNode(node1);
+    printNode(node2);
+    printAnf(anf);
+    printHashMap(anf->hashMap);
+
+    if (generateAnfGraph(anf, graphOutput) != OK) {
+        freeNode(node1);
+        freeNode(node2);
+        freeAnf(anf);
+        return ERROR;  
+    }
+
     if (deleteAllNodesFromAnf(anf, true) != OK) {
         freeNode(node1);
         freeNode(node2);
@@ -681,6 +622,120 @@ int singleFunctionTest(char* graphOutput) {
     return OK;
 }
 
+
+int multipleFunctionsTest(char* graphOutput) {
+
+    tVar varArray[] = {
+        createVar("var1", true),
+        createVar("var2", true),
+        createVar("var3", true)
+    };
+
+    tVar varArray2[] = {
+        createVar("var4", true),
+        createVar("var5", true),
+        createVar("var6", true)
+    };
+
+    tVar varArray3[] = {
+        createVar("var3", false),
+        createVar("var8", true),
+        createVar("var9", true),
+        createVar("var10", true)
+    };
+
+    tVar varArray4[] = {
+        createVar("var11", true),
+        createVar("var12", true),
+        createVar("var13", true)
+    };
+
+    tAnf* anf1 = newAnf(20, 0.7);
+    if (anf1 == NULL) {
+        return ERROR;
+    }
+
+    tNode* node1 = newNodeWithVarsInAnf(anf1, varArray, ARRAY_SIZE(varArray));
+    if (node1 == NULL) {
+        freeAnf(anf1);
+        return ERROR;
+    }
+
+    tNode* node2 = newNodeWithVarsInAnf(anf1, varArray2, ARRAY_SIZE(varArray2));
+    if (node2 == NULL) {
+        freeNode(node1);
+        freeAnf(anf1);
+        return ERROR;
+    }
+
+    printAnf(anf1);
+    printHashMap(anf1->hashMap);
+
+    tAnf* anf2 = newAnf(20, 0.7);
+    if (anf2 == NULL) {
+        freeNode(node1);
+        freeNode(node2);
+        freeAnf(anf1);
+        return ERROR;
+    }
+
+    tNode* node3 = newNodeWithVarsInAnf(anf2, varArray3, ARRAY_SIZE(varArray3));
+    if (node3 == NULL) {
+        freeNode(node1);
+        freeNode(node2);
+        freeAnf(anf1);
+        freeAnf(anf2);
+        return ERROR;
+    }
+
+    tNode* node4 = newNodeWithVarsInAnf(anf2, varArray4, ARRAY_SIZE(varArray4));
+    if (node4 == NULL) {
+        freeNode(node1);
+        freeNode(node2);
+        freeNode(node3);
+        freeAnf(anf1);
+        freeAnf(anf2);
+        return ERROR;
+    }
+
+    printAnf(anf2);
+    printHashMap(anf2->hashMap);
+
+    tAnf* anf3 = newAnfFrom2Anfs(anf1, anf2, true);
+    if (anf3 == NULL) {
+        freeNode(node1);
+        freeNode(node2);
+        freeNode(node3);
+        freeNode(node4);
+        freeAnf(anf1);
+        freeAnf(anf2);
+        return ERROR;
+    }
+
+    printAnf(anf3);
+    printHashMap(anf3->hashMap);
+
+    if (generateAnfGraph(anf3, graphOutput) != OK) {
+        freeNode(node1);
+        freeNode(node2);
+        freeNode(node3);
+        freeNode(node4);
+        freeAnf(anf1);
+        freeAnf(anf2);
+        return ERROR;  
+    }
+
+    freeNode(node1);
+    freeNode(node2);
+    freeNode(node3);
+    freeNode(node4);
+
+    freeAnf(anf1);
+    freeAnf(anf2);
+
+    return OK;   
+}
+
 int main(int argc, char* argv[]) {
 
     struct stat st = {0};
@@ -693,11 +748,13 @@ int main(int argc, char* argv[]) {
     int test1Result = hashMapTest("./testOutput/test1-graph.gv");
     freopen ("./testOutput/test2-output", "a+", stdout);
     int test2Result = singleFunctionTest("./testOutput/test2-graph.gv");
-    
+    freopen ("./testOutput/test3-output", "a+", stdout);
+    int test3Result = multipleFunctionsTest("./testOutput/test3-graph.gv");
 
     freopen("/dev/tty", "w", stdout);
-    printf("TEST 01: HashMap Test         ..... Result: %s %d\n", (test1Result == OK) ? "SUCCESS" : "FAILURE", test1Result);
-    printf("TEST 02: Single Function Test ..... Result: %s %d\n", (test2Result == OK) ? "SUCCESS" : "FAILURE", test2Result);
+    printf("TEST 01: HashMap Test            ..... Result: %s %d\n", (test1Result == OK) ? "SUCCESS" : "FAILURE", test1Result);
+    printf("TEST 02: Single Function Test    ..... Result: %s %d\n", (test2Result == OK) ? "SUCCESS" : "FAILURE", test2Result);
+    printf("TEST 03: Multiple Functions Test ..... Result: %s %d\n", (test3Result == OK) ? "SUCCESS" : "FAILURE", test3Result);
     
 
     // printf("TEST 01: %d\n", emptyAnfs());
